requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
celery>=5.3.0
redis>=5.0.0
mysqlclient>=2.2.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
django-cors-headers>=4.3.0
drf-spectacular>=0.26.0
pytest>=7.4.0
pytest-django>=4.5.5
dj-database-url>=2.1.0
drf-nested-routers>=0.93.4
django-filter>=23.3
django-celery-beat>=2.5.0
python-json-logger>=2.0.7
whitenoise>=6.5.0
factory-boy>=3.3.0
